LayoutTests/editing/undomanager/undoscope-attribute.html

<!doctype html>
<html>
<head>
<title>undoscope attribute</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
This tests element's undoscope attribute.

<div id="outerContainer">
    <div id="innerContainer">
        <div id="element" undoscope></div>
    </div>
</div>

<script>
var outerContainer = document.getElementById("outerContainer");
var innerContainer = document.getElementById("innerContainer");
var element = document.getElementById("element");

function assert_element_has_undomanager() {
    assert_equals(element.undoManager.toString(), "[object UndoManager]");
}

function assert_element_not_have_undomanager() {
    assert_equals(element.undoManager, null);
}

test(function() {
    assert_own_property(element, "undoScope");
    assert_equals(element.undoScope, true);
}, "An element has the undoscope attribute set to true.");

test(function() {
    assert_element_has_undomanager();
}, "element.undoManager returns an UndoManager object.");

test(function() {
    element.undoScope = false;

    assert_element_not_have_undomanager();
}, "After element.undoScope = false, element.undoManager returns null.");

test(function() {
    element.undoScope = true;

    assert_element_has_undomanager();
}, "After element.undoScope = true, element.undoManager returns an UndoManager object.");

test(function() {
    element.setAttribute('undoscope', '');
    element.removeAttribute('undoscope');

    assert_element_not_have_undomanager();
}, "After element.removeAttribute('undoscope'), element.undoManager returns null.");

test(function() {
    element.setAttribute('undoscope', '');

    assert_element_has_undomanager();
}, "After element.setAttribute('undoscope', ''), element.undoManager returns an UndoManager object.");

test(function() {
    element.setAttribute('undoscope', 'undoscope');

    assert_element_has_undomanager();
}, "After element.setAttribute('undoscope', 'undoscope'), element.undoManager returns an UndoManager object.");

test(function() {
    element.setAttribute('undoscope', '');
    element.undoScope = false;

    assert_element_not_have_undomanager();
}, "After element.setAttribute('undoscope', ''); element.undoScope = false;, element.undoManager returns null.");

test(function() {
    element.removeAttribute('undoscope');
    element.undoScope = true;

    assert_element_has_undomanager();
}, "After element.removeAttribute('undoscope'); element.undoScope = true;, element.undoManager returns an UndoManager object.");

test(function() {
    element.undoScope = true;
    element.removeAttribute('undoscope');

    assert_element_not_have_undomanager();
}, "After element.undoScope = true; element.removeAttribute('undoscope');, element.undoManager returns null.");

test(function() {
    element.undoScope = false;
    element.setAttribute('undoscope', '');

    assert_element_has_undomanager();
}, "After element.undoScope = false; element.setAttribute('undoscope', '');, element.undoManager returns an UndoManager object.");

test(function() {
    innerContainer.contentEditable = "true";

    assert_element_not_have_undomanager();
}, "After the parant of element becomes editable, element.undoManager returns null.");

test(function() {
    innerContainer.contentEditable = "false";

    assert_element_has_undomanager();
}, "After the parant of element becomes non-editable, element.undoManager returns an UndoManager object.");

test(function() {
    outerContainer.contentEditable = "true";
    innerContainer.contentEditable = "inherit";

    assert_element_not_have_undomanager();
}, "After the parant of element becomes editable through inheritance, element.undoManager returns null.");

test(function() {
    innerContainer.contentEditable = "false";
    element.contentEdiable = "true";

    assert_element_has_undomanager();
}, "After the parant of element becomes non-editable and element becomes editable, element.undoManager returns an UndoManager object.");

test(function() {
    element.undoManager.transact({"execute": function () { }}, false);
    
    assert_equals(element.undoManager.length, 1);

    element.undoScope = false;
    element.undoScope = true;

    assert_equals(element.undoManager.length, 0);
}, "element.undoManager is disconnected when element.undoScope = false.");

test(function() {
    element.undoManager.transact({"execute": function () { }}, false);
    
    assert_equals(element.undoManager.length, 1);

    innerContainer.contentEditable = "true";
    innerContainer.contentEditable = "false";

    assert_equals(element.undoManager.length, 0);
}, "element.undoManager is disconnected when its parent becomes editable.");

test(function() {
    outerContainer.contentEditable = "false";
    innerContainer.contentEditable = "false";
    element.undoManager.transact({"execute": function () { }}, false);

    assert_equals(element.undoManager.length, 1);

    outerContainer.contentEditable = "true";

    assert_equals(element.undoManager.length, 1);
}, "Setting contentEditable of an ancestor of element to true doesn't disconnect element.undoManager if element is still an editing host.");
</script>
</body>
</html>
